rename files for consistent camelcase
