+++ lib/test/OneWireMockTest.cpp
@@ -22,11 +22,11 @@
 #include "../inc/OneWireMockDevice.h"
 #include "../inc/OneWireMockDriver.h"
 #include "DS18B20.h"
-#include "DS18B20mock.h"
+#include "DS18B20Mock.h"
 #include "DS2408.h"
-#include "DS2408mock.h"
+#include "DS2408Mock.h"
 #include "DS2413.h"
-#include "DS2413mock.h"
+#include "DS2413Mock.h"
 #include "MotorValve.h"
 
 namespace Catch {
